refactor: 요구사항에 명시된 domain 패키지에 존재 해야할 class 이동 시킴. 핵심 비즈니스 로직 당담하는 class와 entity들을 domain 패키지 하위로 이동 시킴.
