+++ src/main/java/racingcar/domain/model/CarMoveRule.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain.model;
 
 public interface CarMoveRule {
     boolean isGo();
