+++ src/main/java/racingcar/domain/model/DefaultCarMoveRule.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain.model;
 
 import racingcar.utils.Random;
