+++ src/main/java/racingcar/domain/model/RacingGame.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain.model;
 
 import java.util.List;
